# Remove the size parameter from Precomputations