Also move some parameters out of Precomputations.

Before:
```
Run on (4 X 3310 MHz CPU s)
CPU Caches:
  L1 Data 32K (x4)
  L1 Instruction 32K (x4)
  L2 Unified 262K (x4)
  L3 Unified 6291K (x1)
------------------------------------------------------------------------------
Benchmark                                       Time           CPU Iterations
------------------------------------------------------------------------------
BM_ComputeGeopotentialCpp/2                185057 ns     183580 ns       3739
BM_ComputeGeopotentialCpp/3                262468 ns     267049 ns       2804
BM_ComputeGeopotentialCpp/5                361499 ns     359818 ns       1951
BM_ComputeGeopotentialCpp/10              1113582 ns    1112307 ns        561
BM_ComputeGeopotentialF90/2                189437 ns     189804 ns       3452
BM_ComputeGeopotentialF90/3                231198 ns     229490 ns       2991
BM_ComputeGeopotentialF90/5                314695 ns     312836 ns       2244
BM_ComputeGeopotentialF90/10               797973 ns     792518 ns        748
BM_ComputeGeopotentialDistance/150000      190063 ns     191924 ns       3739
BM_ComputeGeopotentialDistance/500000       71952 ns      72313 ns      11218
BM_ComputeGeopotentialDistance/5000000      15912 ns      15992 ns      44872
```
After:
```
Run on (4 X 3310 MHz CPU s)
CPU Caches:
  L1 Data 32K (x4)
  L1 Instruction 32K (x4)
  L2 Unified 262K (x4)
  L3 Unified 6291K (x1)
------------------------------------------------------------------------------
Benchmark                                       Time           CPU Iterations
------------------------------------------------------------------------------
BM_ComputeGeopotentialCpp/2                182962 ns     183580 ns       3739
BM_ComputeGeopotentialCpp/3                233103 ns     233668 ns       2804
BM_ComputeGeopotentialCpp/5                359574 ns     359818 ns       1951
BM_ComputeGeopotentialCpp/10               975638 ns     980220 ns        748
BM_ComputeGeopotentialF90/2                191297 ns     187752 ns       3739
BM_ComputeGeopotentialF90/3                252026 ns     250358 ns       2804
BM_ComputeGeopotentialF90/5                313378 ns     312836 ns       2244
BM_ComputeGeopotentialF90/10               794720 ns     800005 ns        897
BM_ComputeGeopotentialDistance/150000      186832 ns     187752 ns       3739
BM_ComputeGeopotentialDistance/500000       71877 ns      70922 ns      11218
BM_ComputeGeopotentialDistance/5000000      15804 ns      15645 ns      44872
```